requirements.txt
requests~=2.27.1
Pillow~=9.1.1
playwright~=1.45.0
pytest~=8.2.2
selenium~=3.141.0